Update bash create_venv.sh
It will check if there's an internet connection by pinging Google's DNS server 8.8.8.8 using the ping command, if it's successful it will prompt the user to install packages from a requirements.txt file, otherwise it will prompt the user to install packages from a local path by specifying the directory path where the packages are located. The script will then use the pip install --no-index --find-links=$local_path -r $requirements_path command to install the packages from the local path.
Please note that this script needs to be run with a superuser or an admin privilege if the packages are not in the system path.